Finish major refactors with layout, included splide library for carousels
