+++ src/layouts/index.jsx
@@ -18,6 +18,16 @@ export const Layout = ({ children, center }) => {
   );
 };
 
-export const Page = ({ children }) => {
-  return <div className="container pl-10 pr-10">{children}</div>;
+export const Page = ({ children, modifier = "" }) => {
+  return <div className={`container pl-10 pr-10 ${modifier}`}>{children}</div>;
+};
+
+export const Grid = ({ children, modifier = "" }) => {
+  return <div className={`grid ${modifier}`}>{children}</div>;
+};
+
+export const Cell = ({ children, size = 12, modifier = "" }) => {
+  return (
+    <div className={`cell cell--size-${size} ${modifier}`}>{children}</div>
+  );
 };
