+++ src/sass/components/_app-navigation.scss
@@ -1,4 +1,5 @@
 .app-navigation {
   background-color: $col-white;
   padding: 4rem 2rem;
+  position: relative;
 }
