+++ src/sass/components/_buttons.scss
@@ -37,7 +37,7 @@
 }
 
 .filter-btn {
-  min-width: 8rem;
+  width: 8rem;
   min-height: 10rem;
   padding: 1.1rem 0.8rem;
   background-color: $col-white;
