+++ src/sass/components/_menu.scss
@@ -1,13 +1,13 @@
 .menu {
   @include flex(column, center, flex-start);
-  position: fixed;
-  left: 0;
+  position: absolute;
+  left: -2rem;
   top: 0;
   z-index: 2;
   background-color: $col-white;
   width: 19rem;
   height: 100vh;
-  transform: translateX(-20rem);
+  transform: translateX(-120%);
   transition: all 0.3s;
   box-shadow: $bs-30;
 
@@ -16,6 +16,10 @@
   }
   button {
     margin-bottom: 2.4rem;
+
+    @media screen and (min-width: 1200px) {
+      display: none;
+    }
   }
 
   &__user {
@@ -41,5 +45,11 @@
         text-decoration: none;
       }
     }
+  }
+
+  @media screen and (min-width: 1200px) {
+    position: static;
+    margin-left: -4rem;
+    height: 100%;
   }
 }
