+++ src/sass/pages/_category.scss
@@ -3,18 +3,10 @@
     margin-bottom: 5rem;
   }
   &__filters {
-    @include flex;
-    @include gap(2rem, 2rem);
-    flex-wrap: wrap;
     margin-bottom: 5rem;
   }
 
   .h2 {
     margin-bottom: 2rem;
-  }
-  &__products {
-    @include flex;
-    @include gap(2rem, 2rem);
-    flex-wrap: wrap;
   }
 }
