+++ src/screens/SignUp/index.jsx
@@ -1,61 +1,33 @@
-import React, { useContext } from "react";
-import Button from "../../components/Button";
-import Form from "../../components/Form";
-import Input, { Field } from "../../components/Inputs";
-import UnauthorizedLayout from "../../layouts/UnauthorizedLayout";
-import { useNavigate } from "react-router-dom";
-import { AuthContext } from "../../context/Auth";
-import Heading from "../../components/Heading";
-import { Logo } from "../../components/Icons";
+import React from "react";
+import LogIn from "../../features/Login";
+import SignUp from "../../features/SignUp";
+import useMediaQuery from "../../hooks/useMediaQuery";
+import { Cell, Grid, Page } from "../../layouts";
 
-const SignUp = () => {
-  const { signUp } = useContext(AuthContext);
-  const navigate = useNavigate();
+const SignUpPage = () => {
+  const isDesktop = useMediaQuery("(min-width: 1200px)");
 
-  const onSubmit = (formData) => {
-    const email = formData.get("email");
-    const password = formData.get("password");
-    const passwordConfirm = formData.get("password-confirm");
-
-    if (!email || !password || !passwordConfirm || password !== passwordConfirm)
-      return;
-    signUp({ email, password });
-  };
-
-  const signUpForm = (
-    <React.Fragment>
-      <div className="auth">
-        <Logo className="logo" />
-        <Heading size="big">Create an account</Heading>
-        <Heading size="small">
-          Welcome friend, enter your details so lets get started in ordering
-          food.
-        </Heading>
-        <Form onSubmit={onSubmit}>
-          <Field label="Email">
-            <Input type="email" name="email" placeholder="Email" />
-          </Field>
-          <Field label="Password">
-            <Input type="password" name="password" placeholder="Password" />
-          </Field>
-          <Field label="Password Confirm">
-            <Input
-              type="password"
-              name="password-confirm"
-              placeholder="Password Confirm"
-            />
-          </Field>
-          <Button long type="submit">
-            Sign up
-          </Button>
-          <Button onClick={() => navigate("/log-in")} long secondary>
-            Go to login
-          </Button>
-        </Form>
-      </div>
-    </React.Fragment>
+  // TODO przemyśl nazwy klas
+  return (
+    <Page>
+      <Grid>
+        {isDesktop ? (
+          <>
+            <Cell size={6} modifier="center-horizontally">
+              <LogIn />
+            </Cell>
+            <Cell size={6} modifier="center-horizontally">
+              <SignUp />
+            </Cell>
+          </>
+        ) : (
+          <Cell modifier={"center-horizontally"}>
+            <SignUp />
+          </Cell>
+        )}
+      </Grid>
+    </Page>
   );
-  return <UnauthorizedLayout firstColumn={signUpForm} />;
 };
 
-export default SignUp;
+export default SignUpPage;
